Revert "VAULT-330 KS-IOF-F-12 Security audit. Declaring partyCount and threshold parameters as uints."
This reverts commit dfc563a374cafe9224d0f9da02738c53c6af779a.

+++ ecdsa/keygen/local_party.go
@@ -142,7 +142,7 @@ func (p *LocalParty) ValidateMessage(msg tss.ParsedMessage) (bool, *tss.Error) {
 		return ok, err
 	}
 	// check that the message's "from index" will fit into the array
-	if maxFromIdx := p.params.PartyCount() - 1; maxFromIdx < uint(msg.GetFrom().Index) {
+	if maxFromIdx := p.params.PartyCount() - 1; maxFromIdx < msg.GetFrom().Index {
 		return false, p.WrapError(fmt.Errorf("received msg with a sender index too great (%d <= %d)",
 			p.params.PartyCount(), msg.GetFrom().Index), msg.GetFrom())
 	}
