+++ eddsa/signing/round_1.go
@@ -89,7 +89,7 @@ func (round *round1) prepare() error {
 	xi := round.key.Xi
 	ks := round.key.Ks
 
-	if round.Threshold()+1 > uint(len(ks)) {
+	if round.Threshold()+1 > len(ks) {
 		// TODO: this should not panic
 		return fmt.Errorf("t+1=%d is not consistent with the key count %d", round.Threshold()+1, len(ks))
 	}
